In this step, I'm installing some tools to clean the code ( phpcsfixer and psalm ).
Also, I create a "scripts" part on composer to ease the launch of this tools and run the tests.

After that, and thanks to apply TDD in the previous step, I separate the code in more objects ( Coordinates, Direction, RoverStatus ) to improve the Single Responsability Principle and use composition.

The next step, let's see if the Rover moves a little with commands

TODO:
- More tests
- Control of the heading direction after each turn
- Control of the Rover if it goes out of the Plateau ( Houston, we have a problem.. )